test: the harness shims against the class they stand in for
Five harnesses replace `zpp::hypervisor::hypervisor` with a stand-in
carrying only the members the code they cut out touches. That is what
makes them possible - the real class is megabytes of per-processor
arrays - and it is also a copy, which is somewhere for a silent
disagreement to live.

Two kinds of drift, and only one announces itself:

- A member the real source uses and a shim lacks is a **compile error**.
  That is the drift that has bitten three times, and it needs no help.
- A member present in both with a *different declaration* compiles
  perfectly and tests the wrong thing.

**It found one of the second kind on the first run.**
`tests/ap_start_up`'s shim declared `std::uint64_t platform_apic_id
[max_cpus]` where the hypervisor has `std::uint32_t`. Every roster walk
in `start_up_broadcast` - which that harness cuts out and drives - was
being tested against an array with twice the stride of the real one.
Nothing failed, because every index was consistent with itself.

Same argument as `start_up_handoff_state`'s constant comparison in
tests/ap_start_up/build.sh, generalised: where a test carries a copy of
something the hypervisor owns, the copy has to be checked against the
original or the test is about the copy.

What it compares, and what it deliberately does not:

- **Bounds always.** A shim array of the wrong rank or extent is the
  failure this exists for.
- **The element type only where it is fundamental.** A shim standing
  `page_table_stub` in for `arch::x86_64::page_table` is the whole point
  of a shim; comparing those spellings would report every harness as
  broken. An integer width is different - it is a stride.
- **A bound the hypervisor writes as a conditional** -
  `[nested_vmx::enabled ? max_cpus : 1]` - has two legitimate values, and
  a harness that always compiles with the feature on is right to carry
  the taken branch. Accepted only if the shim's bound is one of the two,
  so a third is still a disagreement.
- **Members only in a shim are reported, not failed.** Every harness adds
  instrumentation and there is no way to tell that from a typo except by
  reading - so they are printed, which is what makes a typo visible: a
  misspelled member silently becomes harness-only and appears in a list
  that is otherwise short and familiar.

Two things went wrong writing it, both worth keeping:

- The first parser split the file on semicolons and trimmed back from the
  last brace to find where each declaration began. It parsed **47 members
  out of 329**. The self-check that the real header must yield at least a
  hundred is what caught it, and it is a hard failure rather than a
  warning for the reason this whole batch keeps running into: **a checker
  comparing nothing reports success.**
- The second version tracked absolute brace depth to skip nested structs
  and parsed **zero**, because the real header nests the class in two
  namespaces and a shim in one, so "depth 1" means different things in
  the two files. It now anchors on the `class hypervisor` line and counts
  from there. Without the nesting check, a harness-local
  `struct attempt { std::uint64_t vector; }` was compared against the
  hypervisor's own top-level `vector` and reported as a disagreement -
  a false report, which is the failure a checker can least afford.

Verified both directions: fixing the width clears it, and widening the
real declaration to `[max_cpus][2]` makes two harnesses fail naming the
bounds.

+++ tests/ap_start_up/shim/zpp/hypervisor/hypervisor.h
@@ -211,7 +211,7 @@ public:
     // The platform's roster, which is the only thing a broadcast - the
     // form that names no destination - can be resolved against.
     std::size_t number_of_platform_processors{};
-    std::uint64_t platform_apic_id[max_cpus]{};
+    std::uint32_t platform_apic_id[max_cpus]{};
 
     // What the guest asked for, as counters, which is all the log has to
     // say about a command it decided not to act on.
